--- a/src/games/strategy/engine/message/UnifiedMessenger.java
+++ b/src/games/strategy/engine/message/UnifiedMessenger.java
@@ -46,9 +46,7 @@ public class UnifiedMessenger {
   // threads wait on these latches for the hub to return invocations
   // the latch should be removed from the map when you countdown the last result
   // access should be synchronized on m_pendingLock
-  // TODO: how do these get shutdown
-  // when we exit a game or close
-  // triplea?
+  // TODO: how do these get shutdown when we exit a game or close triplea?
   private final Map<GUID, CountDownLatch> m_pendingInvocations = new HashMap<GUID, CountDownLatch>();
   // after the remote has invoked, the results are placed here
   // access should be synchronized on m_pendingLock
@@ -364,7 +362,6 @@ public class UnifiedMessenger {
       stream.println("Remote nodes with implementors:" + m_results);
       stream.println("Remote nodes with implementors:" + m_pendingInvocations);
     }
-  }
 
   @Override
   public String toString() {