Decrease runtime of tests by looking at the top 3 slowest tests. About 10s was cut from total test execution time, now at 21s from 31s (on my system).

To get this gain:
-  excessive wait time was decreased from 400ms to 10ms in one test to gain a lot of time (about 5s). Meanwhile that same test had the concurrency bumped up.
- another unnecessary sleep delay was removed altogether
- Another slow test had thoroughness reduced by lowering the number of test clients, saving a few more seconds
- A do-nothing test method and extra map parsing were removed for almost a 2 second gain


PR - based on: #109

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/triplea-game/triplea/115)
<!-- Reviewable:end -->