@DanVanAtta 2015-11-17
Rebase not quite successful, will fix it up a bit later today.